using macros
swap functions to macro

+++ Operaciones_Matenaticas.h
@@ -1,16 +1,16 @@
 #ifndef OPERACIONES_MATENATICAS_H
 #define OPERACIONES_MATENATICAS_H
 
-float cal_square(float num1);
+double cal_square(double num1);
 //calculate and return square of number
 
-float cal_addition(float num1, float num2);
+double cal_addition(double num1, double num2);
 //calculate and return addition of both
 
-float circle_area(float radius);
+double circle_area(double radius);
 //return the area of cicle
 
-float maximal(float num1, float num2);
+double maximal(double num1, double num2);
 //return the maximun of both
 
 #endif // OPERACIONES_MATENATICAS_H
